app: add launcher for spp_nfv and spp_vm
spp-nfv.sh is for launching spp_nfv which uses resource on host and
does not use vhost interface. It mounts /dev/hugepages and /var/run
to share them with primary process. spp_nfv runs as a secondary.

On the other hand, spp-vm.sh is for launching spp_vm using vhost
interface. It also mounts /dev/hugepages and /var/run but cannot
access to host resource. spp_vm runs as a primary actually.

Signed-off-by: Yasufumi Ogawa <ogawa.yasufumi@lab.ntt.co.jp>