build: fix bug for proxy not affect
docker requires to configure proxy setting if you are in proxy
network.

Proxy settings for docker daemon is described in config files
contained in /etc/systemd/system/docker.service.d/.
  * docker.conf : define DNS address
  * http-proxy.conf : define proxy address

In addition, you need to assing proxy params with `--build-arg'
explicitly when you run it.

This update is for adding '--build-arg' in 'build.sh'.

Signed-off-by: Yasufumi Ogawa <ogawa.yasufumi@lab.ntt.co.jp>