project: change directory structure
Spp-container is a trial project and test files are mixed with spp app
scripts. Now bug in spp-nvf.sh is fixed and it is the time to clean up
project.

Test script file, which are for running dpdk apps on host and
containers, is moved to experimental/.

Some of scripts referring external files are needed to be corrected its
path. It is the next todo.

Signed-off-by: Yasufumi Ogawa <ogawa.yasufumi@lab.ntt.co.jp>