build: change build script to python
To give detailed options, change build script from sh to python. For
this update, 'env.sh' is removed and environmental variables are moved
to 'app/conf.py'. Although 'build/env.sh' is still required for
'build/run.sh', but it is changed to be automatically generated from
'build/build.py' and you don't need to care about it.

Signed-off-by: Yasufumi Ogawa <ogawa.yasufumi@lab.ntt.co.jp>